[ICU-7097] Enabled java compiler option -Xlint:all,-deprecation,-dep-ann for API projects.  Fixed redundant cast warnings detected by javac in various files.  Added @SuppressWarnings for intended usage of fallthrough in switch statements.  Enabled fallthrough warning in Eclipse project files. X-Commit-URL: https://ssl.icu-project.org/trac/changeset/26568 Former-commit-id: 71fa6bff53fba46267c40b573f19174473983aa6

+++ icu4j/main/classes/charset/src/com/ibm/icu/charset/CharsetCallback.java
@@ -382,7 +382,7 @@ public class CharsetCallback {
         char temp;
         
         do {
-            digit = (int)(i % radix);
+            digit = i % radix;
             buffer[sourceIndex + length++] = (char)(digit <= 9 ? (0x0030+digit) : (0x0030+digit+7));
             i = i/radix;
         } while (i != 0 && (sourceIndex + length) < buffer.length);
